requirements.txt
Flask==3.0.3
Pillow==10.4.0
Werkzeug==3.0.4
gunicorn
load_dotenv
azure-storage-blob
